Restructure test
